fixed review

+++ contracts/helpers/TestCoinContract.sol
@@ -3,13 +3,13 @@ pragma solidity >=0.8.0 <0.9.0;
 import "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 
 contract TestCoin is ERC20 {
-    uint currentSupply;
+    uint256 currentSupply;
     
-    constructor(uint _initialSupply) ERC20("TestCoin", "TC") { 
+    constructor(uint256 _initialSupply) ERC20("TestCoin", "TC") { 
         currentSupply = _initialSupply;
     }
 
-    function mint(uint amount) external {
+    function mint(uint256 amount) external {
         require(currentSupply > 0);
         _mint(msg.sender, amount);
         currentSupply -= amount;
